+++ Develop/db/db.json
@@ -1,1 +1,7 @@
-[]
+[
+    {
+        "title": "Groceries",
+        "text": "Get Milk",
+        "id": "b796609b-2142-4531-a565-fe558db1a7b4"
+    }
+]
