+++ Develop/server.js
@@ -1,7 +1,7 @@
 //Dependencies for express and the built in path functionality
 const express = require('express');
-const path = require('path');
-const api = require('./routes/index.js');
+const api = require('./routes/api/apiRoutesIndex.js');
+const htmlRouter = require('./routes/htmlRoutes');
 
 const PORT = process.env.port || 3001;
 const app = express();
@@ -9,25 +9,11 @@ const app = express();
 // Middleware for parsing JSON and urlencoded form data
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
-app.use('/api', api);
-
 // Serves static files in public folder
 app.use(express.static('public'));
+app.use('/api', api);
+app.use('/', htmlRouter);
 
-// GET Route for homepage
-app.get('/', (req, res) =>
-  res.sendFile(path.join(__dirname, '/public/index.html'))
-);
-
-// GET Route for notes page
-app.get('/notes', (req, res) =>
-  res.sendFile(path.join(__dirname, '/public/notes.html'))
-);
-
-// Wildcard route to direct any unspecified path back to index.html
-app.get('*', (req, res) =>
-  res.sendFile(path.join(__dirname, '/public/index.html'))
-);
 
 // Notify in console of port app is listening on.
 app.listen(PORT, () =>
